[REFACTOR] Solid substitution of SmartRecyclerAdapter as interface with implementations
